PythonWorker/requirements.txt
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
pandas==2.2.0
scikit-learn==1.4.0
